Add record details component and associated routing.

+++ src/app/app.component.ts
@@ -60,26 +60,28 @@ export class AppComponent implements OnInit {
     this.appStateService.setConfig(this.config);
   }
 
-  onActivate(componentReference: RecordslistComponent) {
-    console.log(componentReference);
-    componentReference.saveRecordRequested.subscribe((record: Record) => {
-      this.onSaveRecordRequested(record);
-    });
-    componentReference.deleteRecordRequested.subscribe((recordDeletionID: RecordDeletionID) => {
-      this.onDeleteRecordRequested(recordDeletionID);
-    });
-    componentReference.updateRecordRequested.subscribe((record: Record) => {
-      this.onUpdateRecordRequested(record);
-    });
-    componentReference.addKeywordsRequested.subscribe((records: Record[]) => {
-      this.onAddKeywordsRequested(records);
-    });
-    componentReference.pageChanged.subscribe((newPage: number) => {
-      this.onPageChanged(newPage);
-    });
-    componentReference.sortRequested.subscribe((sortRequest: [string, boolean]) => {
-      this.onSortRequested(sortRequest);
-    });
+  onActivate(componentReference: any) {
+    if (componentReference instanceof RecordslistComponent) {
+      console.log(componentReference);
+      componentReference.saveRecordRequested.subscribe((record: Record) => {
+        this.onSaveRecordRequested(record);
+      });
+      componentReference.deleteRecordRequested.subscribe((recordDeletionID: RecordDeletionID) => {
+        this.onDeleteRecordRequested(recordDeletionID);
+      });
+      componentReference.updateRecordRequested.subscribe((record: Record) => {
+        this.onUpdateRecordRequested(record);
+      });
+      componentReference.addKeywordsRequested.subscribe((records: Record[]) => {
+        this.onAddKeywordsRequested(records);
+      });
+      componentReference.pageChanged.subscribe((newPage: number) => {
+        this.onPageChanged(newPage);
+      });
+      componentReference.sortRequested.subscribe((sortRequest: [string, boolean]) => {
+        this.onSortRequested(sortRequest);
+      });
+    }
  }
 
   /**
@@ -114,7 +116,7 @@ export class AppComponent implements OnInit {
         } else {
           this.setRecords(res);
         }
-        if ((this.lastSearchRequest.Sort !== null) && (this.recordListComponent !== null)) {
+        if ((this.lastSearchRequest.Sort !== null) && (this.recordListComponent !== undefined) && (this.recordListComponent !== null)) {
           this.recordListComponent.updateSortOptionsFromSortId(this.lastSearchRequest.Sort);
         }
       }, err => {
